the frame is framier than before

+++ src/main/java/arathain/vigorem/anim/Keyframe.java
@@ -11,8 +11,8 @@ public final class Keyframe {
 	public final Vec3f scale;
 	public final boolean override;
 
-	public final int frame;
-	public Keyframe(Easing easing, Vec3f translation, Vec3f rot, Vec3f scale, Vec3f offset, int frame) {
+	public final float frame;
+	public Keyframe(Easing easing, Vec3f translation, Vec3f rot, Vec3f scale, Vec3f offset, float frame) {
 		this.override = true;
 		this.easing = easing;
 		this.translation = translation;
@@ -21,7 +21,7 @@ public final class Keyframe {
 		this.frame = frame;
 		this.offset = offset;
 	}
-	public Keyframe(Easing easing, Vec3f translation, Vec3f rot, Vec3f scale, Vec3f offset, int frame, boolean override) {
+	public Keyframe(Easing easing, Vec3f translation, Vec3f rot, Vec3f scale, Vec3f offset, float frame, boolean override) {
 		this.override = override;
 		this.easing = easing;
 		this.translation = translation;
